+++ lib/options/Options.d.ts
@@ -3,8 +3,8 @@
  */
 export default class Options {
     /**
-     * Indicates whether text nodes are displayed on the canvas after they are created or
-     * if they have to be displayed manually.
+     * Indicates whether text nodes are displayed on the canvas after they are created or if they have to
+     * be displayed manually.
      *
      * @property {boolean}
      *
@@ -12,8 +12,15 @@ export default class Options {
      */
     autoDisplay: boolean;
     /**
+     * Class names to add to each text node created.
+     *
+     * @property {Array<string>}
+     */
+    classes: Array<string>;
+    /**
      * @param {Object} options
      * @param {boolean} [options.autoDisplay=true] Indicates whether text nodes are displayed on the canvas after they are created or if they have to be displayed manually.
+     * @param {Array<string>} [options.classes=[]] Class names to add to each text node created.
      */
     constructor(options: Object);
 }