+++ src/options/Options.ts
@@ -6,8 +6,8 @@
 export default class Options {
 
   /**
-   * Indicates whether text nodes are displayed on the canvas after they are created or
-   * if they have to be displayed manually.
+   * Indicates whether text nodes are displayed on the canvas after they are created or if they have to 
+   * be displayed manually.
    * 
    * @property {boolean}
    * 
@@ -16,8 +16,16 @@ export default class Options {
   autoDisplay: boolean = true;
 
   /**
+   * Class names to add to each text node created.
+   * 
+   * @property {Array<string>}
+   */
+  classes: Array<string> = [];
+
+  /**
    * @param {Object} options
    * @param {boolean} [options.autoDisplay=true] Indicates whether text nodes are displayed on the canvas after they are created or if they have to be displayed manually.
+   * @param {Array<string>} [options.classes=[]] Class names to add to each text node created.
    */
   constructor(options: Object) {