+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+0.2.0 / 2019-12-06
+==================
+* [FEATURE] Added an option that can be passed to the gamewriter instance to add class names to each text node created.
+* [FEATURE] Added an option to add an id and classes to each individual text node.
+* [MISC] Updated dependencies to their latest versions.
+* [MISC] Removed unused imports in the Text module.
+
 0.1.0 / 2019-11-24
 ==================
 * Initial commit